Do not modify cpp folder when building

+++ cpp/controllers/controller.cpp
@@ -592,7 +592,7 @@ void Controller::ParseMessage()
 
 void Controller::RejectExtendedMessage()
 {
-    if (msg_bytes.size() < 3 || !msg_bytes[1] || msg_bytes.size() < msg_bytes[1] + 2) {
+    if (msg_bytes.size() < 3 || !msg_bytes[1] || msg_bytes.size() < static_cast<size_t>(msg_bytes[1] + 2)) {
         LogWarn("Truncated extended message");
     }
     else {
